style: main의 레이싱게임 구조체를 새로운 파일을 생성하고 이동시킴
main은 빈 파일이 되어 target에서 tests 언체크 함
